+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Add `IdentityProvider` resource and `Account.identityProvider`
 - Add `EnvelopeEndpoint.clone`
 - Add `EnvelopeProperties.declineAllowed` and `EnvelopeProperties.declineReasonRequired`
+- Add `EnvelopeRecipient.declineReason`
 
 ## [1.5.0] - 2021-11-09
 ### Added